added option of selecting roster, history page
index.html (home) - uncommented the link for history

roster\index.html - put the line '_gaq.push(['_setAccount',
'UA-34842173-1']);' on one line, as this was drawing errors.
Uncommented the history link
Added a select drop down called 'rosterSelect' with the rosters as
options. The options' values are labeled as the first letter of the
season and the year it took place (S12 for Spring 12). Each season now
has a table and title id, with the first letter of the season, the year
it took place, and either table or title (S12table for Spring 12 table).
Added jquery to handle 'rosterSelect's onChange() event. The array
seasons holds the season ids, and the jquery .change(function () {})
handles the onChange. It iterates through seasons, and if the current
season is the same as the selection from rosterSelect, that season's
title and table is shown. Otherwise, that season's table and title are
hidden.

schedule\index.html - uncommented the history link

history\index.html - the history of the KFalcs, through ~1.5 seasons!\

NOTE - Nearly everything here will have to be merged with the actual
file used currently on the site, as a lot of things, especially the
roster page and schedule page, are outdated. I assume you've done merges
before, but just in case if not, kdiff is a great merge tool.